framework/requirements.txt
apscheduler==3.11.2
sqlalchemy==2.0.48
psycopg2-binary==2.9.10
pydantic-settings==2.8.2
python-dotenv==1.1.0

# API REST
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Autenticação JWT
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Validação de e-mail
email-validator==2.2.0
